chore: refactor theme management; migrate to zustand for state management and remove obsolete context
