+++ src/options/index.tsx
@@ -1,17 +1,16 @@
 import { TooltipProvider } from "@/components/ui/tooltip"
-import { ThemeProvider } from "@/context/them-provider-context"
-import { STORAGE_KEYS } from "@/lib/constants"
 import OllamaOptions from "@/options/components/ollama-options"
 
 import "../globals.css"
 
+import { useThemeWatcher } from "@/hooks/use-theme-watcher"
+
 function OptionsIndex() {
+  useThemeWatcher()
   return (
-    <ThemeProvider storageKey={STORAGE_KEYS.THEME.PREFERENCE}>
-      <TooltipProvider>
-        <OllamaOptions />
-      </TooltipProvider>
-    </ThemeProvider>
+    <TooltipProvider>
+      <OllamaOptions />
+    </TooltipProvider>
   )
 }
